Add reusable expanded image hook and modal
