+++ content/webapp/components/ImageEndpointSearchResults/index.tsx
@@ -1,10 +1,4 @@
-import {
-  FunctionComponent,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react';
+import { FunctionComponent, useMemo } from 'react';
 import PhotoAlbum, {
   RenderPhotoProps,
   RenderRowContainer,
@@ -12,15 +6,13 @@ import PhotoAlbum, {
 import styled from 'styled-components';
 
 import { useAppContext } from '@weco/common/contexts/AppContext';
-import { ServerDataContext } from '@weco/common/server-data/Context';
-import Modal from '@weco/common/views/components/Modal';
 import PlainList from '@weco/common/views/components/styled/PlainList';
 import Space from '@weco/common/views/components/styled/Space';
-import ExpandedImage from '@weco/content/components/ExpandedImage';
 import ImageCard from '@weco/content/components/ImageCard';
-import { getImage } from '@weco/content/services/wellcome/catalogue/images';
 import { Image } from '@weco/content/services/wellcome/catalogue/types';
 import { hexToRgb } from '@weco/content/utils/convert-colors';
+import useExpandedImage from './useExpandedImage';
+import ExpandedImageModal from './ExpandedImageModal';
 
 type Props = {
   images: Image[];
@@ -40,7 +32,8 @@ const imageMargin = 16;
 const GalleryContainer = styled.div`
   margin: 0 -${imageMargin / 2}px;
 
-  /* This div is rendered by Gallery so we're targetting it on mobile only to ensure images are aligned on both sides */
+  /* This div is rendered by Gallery so we're targeting it on mobile only to ensure images are aligned on both sides */
+
   .react-photo-gallery--gallery > div {
     justify-content: space-around;
   }
@@ -78,65 +71,7 @@ const ImageEndpointSearchResults: FunctionComponent<Props> = ({
   targetRowHeight,
 }: Props) => {
   const { isFullSupportBrowser } = useAppContext();
-  const [expandedImage, setExpandedImage] = useState<Image | undefined>();
-  const [isActive, setIsActive] = useState(false);
-  const { toggles } = useContext(ServerDataContext);
-
-  const imageMap = useMemo<Record<string, Image>>(
-    () => images.reduce((a, image) => ({ ...a, [image.id]: image }), {}),
-    [images]
-  );
-
-  const setImageIdInURL = (id: string) => {
-    window.location.hash = id;
-  };
-
-  useEffect(() => {
-    const onHashChanged = async () => {
-      // to trim the '#' symbol
-      const hash = window.location.hash.slice(1);
-      if (!hash) {
-        setIsActive(false);
-        setExpandedImage(undefined);
-      }
-      // see if the new hash is already in the imageMap
-      if (imageMap[hash]) {
-        // if it is, update the expanded image
-        setExpandedImage(imageMap[hash]);
-        setIsActive(true);
-      } else {
-        // if it's not, fetch the image and then update
-        const { image } = await getImage({ id: hash, toggles });
-        if (image.type === 'Image') {
-          imageMap[image.id] = image;
-          setExpandedImage(image);
-        } else if (image.type === 'Error') {
-          setExpandedImage(undefined);
-          setIsActive(false);
-        }
-      }
-    };
-    onHashChanged();
-    window.addEventListener('hashchange', onHashChanged);
-    return () => {
-      window.removeEventListener('hashchange', onHashChanged);
-    };
-  }, [imageMap]);
-
-  useEffect(() => {
-    if (isActive && expandedImage !== undefined) {
-      setImageIdInURL(expandedImage?.id || '');
-    } else {
-      // clear the url of the fragments and also removes the # symbol
-      setImageIdInURL('');
-    }
-  }, [isActive, expandedImage]);
-
-  // In the case that the modal changes the expanded image to
-  // be one that isn't on this results page, this index will be -1
-  const expandedImagePosition = images.findIndex(
-    img => expandedImage && img.id === expandedImage.id
-  );
+  const [expandedImage, setExpandedImage] = useExpandedImage(images);
 
   // If there's only two images or less, display them differently (not worth loading the gallery + they display too large)
   const isSmallGallery = images.length < 3;
@@ -177,7 +112,6 @@ const ImageEndpointSearchResults: FunctionComponent<Props> = ({
               onClick={event => {
                 event.preventDefault();
                 setExpandedImage(photo);
-                setIsActive(true);
               }}
               layout="fixed"
               background={
@@ -228,7 +162,6 @@ const ImageEndpointSearchResults: FunctionComponent<Props> = ({
                       if (isSmallGallery) {
                         event.preventDefault();
                         setExpandedImage(result);
-                        setIsActive(true);
                       }
                     }}
                     layout="fixed"
@@ -239,20 +172,11 @@ const ImageEndpointSearchResults: FunctionComponent<Props> = ({
           </ImageCardList>
         </div>
       )}
-      <Modal
-        id="expanded-image-dialog"
-        dataTestId="image-modal"
-        isActive={isActive}
-        setIsActive={setIsActive}
-        maxWidth="1020px"
-      >
-        <ExpandedImage
-          resultPosition={expandedImagePosition}
-          image={expandedImage}
-          isActive={isActive}
-          setExpandedImage={setExpandedImage}
-        />
-      </Modal>
+      <ExpandedImageModal
+        images={images}
+        expandedImage={expandedImage}
+        setExpandedImage={setExpandedImage}
+      />
     </>
   );
 };
